server/requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==25.4.0
boto3==1.40.61
botocore==1.40.61
brotli==1.2.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
croniter==5.0.1
cryptography==46.0.5
cssselect2==0.9.0
defusedxml==0.7.1
distro==1.9.0
Django==6.0.2
django-extensions==4.1
dnspython==2.8.0
docstring_parser==0.17.0
dramatiq==2.0.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.129.2
fastapi-limiter==0.1.6
fonttools==4.62.1
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
honcho==2.0.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
jiter==0.13.0
jmespath==1.1.0
lxml==6.0.2
Markdown==3.7
multidict==6.7.1
numpy==2.4.2
openai==2.21.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-dbapi==0.60b1
opentelemetry-instrumentation-django==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-psycopg2==0.60b1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-wsgi==0.60b1
opentelemetry-instrumentor-dramatiq==0.1.2
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==26.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.11.0
pypdf==5.9.0
pypdfium2==5.6.0
pyphen==0.17.2
pyrate-limiter==4.0.2
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-magic==0.4.27
python-multipart==0.0.22
pytz==2025.2
qdrant-client==1.17.0
redis==7.2.0
regex==2026.2.28
requests==2.32.5
ruff==0.15.2
s3transfer==0.14.0
sentry-sdk==2.53.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
starlette==0.52.1
stripe==14.3.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
weasyprint==63.1
webencodings==0.5.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zopfli==0.4.1
